SAA-334: bookingId should be a long to match precision in NOMIS

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsactivitiesmanagementapi/entity/AppointmentOccurrenceAllocation.kt
@@ -22,7 +22,7 @@ data class AppointmentOccurrenceAllocation(
 
   val prisonerNumber: String,
 
-  val bookingId: Int
+  val bookingId: Long
 ) {
   fun toModel() = AppointmentOccurrenceAllocationModel(
     id = appointmentOccurrenceAllocationId,
